refactor: 重构项目结构，尝试使用rye作项目管理
