+++ .gitignore
@@ -1,9 +1,15 @@
+# python generated files
 __pycache__/
-*.ipynb
-*.csv
-log/
+*.py[oc]
+build/
+dist/
+wheels/
+*.egg-info
+
+# venv
+.venv
+
+# userdata
 data/
-.pytest_cache
-.vscode
-config.*
-checkpoint_time.txt
+log/
+config.yaml
