+++ requirements-dev.lock
@@ -1,0 +1,103 @@
+# generated by rye
+# use `rye lock` or `rye sync` to update this lockfile
+#
+# last locked with the following flags:
+#   pre: false
+#   features: []
+#   all-features: false
+#   with-sources: false
+
+-e file:.
+aiohttp==3.9.3
+    # via aioresponses
+    # via alist-mikananirss
+aioresponses==0.7.6
+aiosignal==1.3.1
+    # via aiohttp
+annotated-types==0.6.0
+    # via pydantic
+anyio==4.3.0
+    # via httpx
+    # via openai
+asyncio==3.4.3
+    # via alist-mikananirss
+attrs==23.2.0
+    # via aiohttp
+beautifulsoup4==4.12.3
+    # via alist-mikananirss
+black==24.3.0
+certifi==2024.2.2
+    # via httpcore
+    # via httpx
+click==8.1.7
+    # via black
+colorama==0.4.6
+    # via click
+    # via loguru
+    # via pytest
+    # via tqdm
+distro==1.9.0
+    # via openai
+feedparser==6.0.11
+    # via alist-mikananirss
+frozenlist==1.4.1
+    # via aiohttp
+    # via aiosignal
+h11==0.14.0
+    # via httpcore
+httpcore==1.0.5
+    # via httpx
+httpx==0.27.0
+    # via openai
+idna==3.6
+    # via anyio
+    # via httpx
+    # via yarl
+iniconfig==2.0.0
+    # via pytest
+loguru==0.7.2
+    # via alist-mikananirss
+multidict==6.0.5
+    # via aiohttp
+    # via yarl
+mypy-extensions==1.0.0
+    # via black
+openai==1.7.0
+    # via alist-mikananirss
+packaging==24.0
+    # via black
+    # via pytest
+pathspec==0.12.1
+    # via black
+platformdirs==4.2.0
+    # via black
+pluggy==1.4.0
+    # via pytest
+pydantic==2.6.4
+    # via openai
+pydantic-core==2.16.3
+    # via pydantic
+pytest==8.1.1
+    # via pytest-asyncio
+pytest-asyncio==0.23.6
+pyyaml==6.0.1
+    # via alist-mikananirss
+ruff==0.3.5
+sgmllib3k==1.0.0
+    # via feedparser
+sniffio==1.3.1
+    # via anyio
+    # via httpx
+    # via openai
+soupsieve==2.5
+    # via beautifulsoup4
+tqdm==4.66.2
+    # via openai
+typing-extensions==4.11.0
+    # via openai
+    # via pydantic
+    # via pydantic-core
+win32-setctime==1.1.0
+    # via loguru
+yarl==1.9.4
+    # via aiohttp
