+++ requirements.lock
@@ -1,0 +1,79 @@
+# generated by rye
+# use `rye lock` or `rye sync` to update this lockfile
+#
+# last locked with the following flags:
+#   pre: false
+#   features: []
+#   all-features: false
+#   with-sources: false
+
+-e file:.
+aiohttp==3.9.3
+    # via alist-mikananirss
+aiosignal==1.3.1
+    # via aiohttp
+annotated-types==0.6.0
+    # via pydantic
+anyio==4.3.0
+    # via httpx
+    # via openai
+asyncio==3.4.3
+    # via alist-mikananirss
+attrs==23.2.0
+    # via aiohttp
+beautifulsoup4==4.12.3
+    # via alist-mikananirss
+certifi==2024.2.2
+    # via httpcore
+    # via httpx
+colorama==0.4.6
+    # via loguru
+    # via tqdm
+distro==1.9.0
+    # via openai
+feedparser==6.0.11
+    # via alist-mikananirss
+frozenlist==1.4.1
+    # via aiohttp
+    # via aiosignal
+h11==0.14.0
+    # via httpcore
+httpcore==1.0.5
+    # via httpx
+httpx==0.27.0
+    # via openai
+idna==3.6
+    # via anyio
+    # via httpx
+    # via yarl
+loguru==0.7.2
+    # via alist-mikananirss
+multidict==6.0.5
+    # via aiohttp
+    # via yarl
+openai==1.7.0
+    # via alist-mikananirss
+pydantic==2.6.4
+    # via openai
+pydantic-core==2.16.3
+    # via pydantic
+pyyaml==6.0.1
+    # via alist-mikananirss
+sgmllib3k==1.0.0
+    # via feedparser
+sniffio==1.3.1
+    # via anyio
+    # via httpx
+    # via openai
+soupsieve==2.5
+    # via beautifulsoup4
+tqdm==4.66.2
+    # via openai
+typing-extensions==4.11.0
+    # via openai
+    # via pydantic
+    # via pydantic-core
+win32-setctime==1.1.0
+    # via loguru
+yarl==1.9.4
+    # via aiohttp
