--- requirements.txt
@@ -1,7 +1,0 @@
-aiohttp>=3.9.1
-asyncio>=3.4.3
-feedparser>=6.0.11
-beautifulsoup4>=4.12.2
-openai==1.7.0
-pyyaml>=6.0.1
-loguru
