--- requirements_dev.txt
@@ -1,6 +1,0 @@
--r requirements.txt
-pytest
-pytest-asyncio
-aioresponses
-ruff
-black
