+++ src/alist_mikananirss/__main__.py
@@ -1,0 +1,5 @@
+import sys
+
+import alist_mikananirss
+
+sys.exit(alist_mikananirss.main())
