--- tests/test_upload.txt
@@ -1,1 +1,0 @@
-test
